Changement Validation

+++ src/main/java/com/example/demo/validations/ValidProduit.java
@@ -5,7 +5,7 @@ public class ValidProduit {
 		return titre != null;
 	}
 	
-	public boolean verifPrix(int unit,int vente) {
+	public boolean verifPrix(float unit, float vente) {
 		return vente > unit;	
 	}
